test(supervisor): move live-listener fixtures out of inline cfg(test) modules
The architecture.yml invariant scan greps every non-allowlisted Rust
file under crates/ + src/ for server-binding patterns (TCP listeners,
axum::serve, etc.) and exits nonzero if any appear. The exclusion glob
`!**/tests/**` only filters files under `tests/` directories — it does
NOT skip inline `#[cfg(test)] mod tests { ... }` modules inside
production source files.

Plan 65 introduced loopback HTTP fixtures (`spawn_one_shot_server` +
the W1/W3 hardening tests) inside the inline test modules of
`tools/web_fetch.rs` and `tools/http_hardening.rs`. Those bindings
tripped the invariant, even though they're test-only.

Move them to integration tests under `crates/mvm-supervisor/tests/`:

- `tests/web_fetch_loopback.rs` — W1 (no auto-redirect) + W3 (exact
  body cap) + W3 boundary, exercising `ReqwestHttpFetcher::fetch`
  against a one-shot 127.0.0.1 server.
- `tests/http_hardening_loopback.rs` — `read_capped`'s chunk-loop
  accumulator under-cap / over-cap / boundary, against the same
  fixture pattern with a bare `reqwest::Client`.

To make the W1/W3 tests reachable from `tests/`, the `test_unsafe_no_ssrf`
seam on `ReqwestHttpFetcher` is promoted from `pub(crate) #[cfg(test)]`
to `pub #[doc(hidden)]`. The name still screams danger; the docstring
spells out why it leaked into the public-API surface (rustdoc hides it).

The SSRF guard literal-IP rejection tests stay inline — they don't
bind sockets and one of them (`ssrf_guard_skipped_when_test_seam_used`)
asserts on the private `enforce_ssrf` field.

CI failure this fixes:

  Architectural invariant #1 violated: a new server appeared in mvm
  outside the substrate allowlist.
  crates/mvm-supervisor/src/tools/web_fetch.rs:        let listener = ...
  crates/mvm-supervisor/src/tools/http_hardening.rs:        let listener = ...
